Consolidate shared frontend UI primitives
