+++ src/components/AnalyticsErrorBoundary.tsx
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { Component, type ReactNode, type ErrorInfo } from "react";
+import { TriangleAlert } from "lucide-react";
 import { analytics } from "../lib/analytics";
 
 interface Props {
@@ -71,19 +72,10 @@ export class AnalyticsErrorBoundary extends Component<Props, State> {
         <div className="min-h-screen flex items-center justify-center bg-gray-50 dark:bg-gray-900 px-4">
           <div className="max-w-md w-full bg-white dark:bg-gray-800 shadow-lg rounded-lg p-6">
             <div className="flex items-center gap-3 mb-4">
-              <svg
+              <TriangleAlert
                 className="h-6 w-6 text-red-500"
-                fill="none"
-                viewBox="0 0 24 24"
-                strokeWidth={2}
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126zM12 15.75h.007v.008H12v-.008z"
-                />
-              </svg>
+                aria-hidden="true"
+              />
               <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-100">
                 Something went wrong
               </h2>
