--- src/pages/Auth/ui-lite.tsx
@@ -1,254 +1,0 @@
-// SPDX-FileCopyrightText: 2026 SecPal
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-import type {
-  ButtonHTMLAttributes,
-  ComponentPropsWithoutRef,
-  FormHTMLAttributes,
-  InputHTMLAttributes,
-  LabelHTMLAttributes,
-  ReactNode,
-} from "react";
-import { cn } from "../../lib/utils";
-
-export function LoginShell({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"main">) {
-  return (
-    <main
-      className={cn(
-        "relative flex min-h-[var(--app-shell-min-height)] flex-col items-center overflow-x-clip bg-white px-6 pt-[calc(1.5rem+var(--app-safe-area-inset-top))] pb-6 text-zinc-950 dark:bg-zinc-950 dark:text-zinc-50 md:px-10 md:pt-[calc(2.5rem+var(--app-safe-area-inset-top))] md:pb-10",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export function LoginCard({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"section">) {
-  return (
-    <section
-      className={cn(
-        "w-full max-w-sm text-zinc-950 dark:text-zinc-50",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export function LoginCardHeader({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"div">) {
-  return (
-    <div
-      className={cn("flex flex-col items-center gap-2 text-center", className)}
-      {...props}
-    />
-  );
-}
-
-export function LoginCardTitle({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"h1">) {
-  return (
-    <h1
-      className={cn(
-        "text-xl font-bold text-zinc-950 dark:text-zinc-50",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export function LoginForm({
-  className,
-  ...props
-}: FormHTMLAttributes<HTMLFormElement>) {
-  return <form className={cn("flex flex-col gap-6", className)} {...props} />;
-}
-
-export function LoginField({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"div">) {
-  return <div className={cn("space-y-2", className)} {...props} />;
-}
-
-export function LoginFieldGroup({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"div">) {
-  return <div className={cn("space-y-6", className)} {...props} />;
-}
-
-export function LoginFieldLabel({
-  className,
-  ...props
-}: LabelHTMLAttributes<HTMLLabelElement>) {
-  return (
-    <label
-      className={cn(
-        "block text-sm font-medium text-zinc-950 dark:text-zinc-100",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export function LoginFieldDescription({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"p">) {
-  return (
-    <p
-      className={cn(
-        "text-sm break-words text-zinc-600 dark:text-zinc-300",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export function LoginFieldError({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"p">) {
-  return (
-    <p
-      className={cn(
-        "text-sm break-words text-red-600 dark:text-red-400",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export function LoginInput({
-  className,
-  ...props
-}: InputHTMLAttributes<HTMLInputElement>) {
-  return (
-    <input
-      className={cn(
-        "flex h-10 w-full rounded-md border border-zinc-300 bg-white px-3 py-2 text-sm text-zinc-950 shadow-xs outline-none transition focus-visible:border-blue-600 focus-visible:ring-2 focus-visible:ring-blue-600/20 disabled:cursor-not-allowed disabled:bg-zinc-100 disabled:text-zinc-500 dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-50 dark:focus-visible:border-blue-500 dark:focus-visible:ring-blue-500/20 dark:disabled:bg-zinc-900 dark:disabled:text-zinc-400",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export type LoginButtonVariant = "default" | "outline";
-
-export function LoginButton({
-  className,
-  variant = "default",
-  type = "button",
-  ...props
-}: ButtonHTMLAttributes<HTMLButtonElement> & {
-  variant?: LoginButtonVariant;
-}) {
-  return (
-    <button
-      type={type}
-      className={cn(
-        "inline-flex h-10 items-center justify-center gap-2 rounded-md px-4 text-sm font-medium outline-none transition focus-visible:ring-2 focus-visible:ring-blue-600/30 disabled:cursor-not-allowed disabled:opacity-50",
-        variant === "outline"
-          ? "border border-zinc-300 bg-white text-zinc-950 hover:bg-zinc-50 dark:border-zinc-700 dark:bg-zinc-950 dark:text-zinc-50 dark:hover:bg-zinc-900"
-          : "bg-zinc-950 text-white hover:bg-zinc-800 dark:bg-white dark:text-zinc-950 dark:hover:bg-zinc-200",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export function LoginFieldSeparator({
-  className,
-  children,
-  ...props
-}: ComponentPropsWithoutRef<"div">) {
-  return (
-    <div
-      role="separator"
-      className={cn(
-        "relative my-2 h-5 text-center text-sm text-zinc-500 dark:text-zinc-400",
-        className
-      )}
-      {...props}
-    >
-      <div
-        aria-hidden="true"
-        className="absolute inset-x-0 top-1/2 h-px -translate-y-1/2 bg-zinc-200 dark:bg-zinc-800"
-      />
-      {children ? (
-        <span className="relative mx-auto inline-block bg-white px-2 dark:bg-zinc-950">
-          {children}
-        </span>
-      ) : null}
-    </div>
-  );
-}
-
-export function LoginStatusMessage({
-  className,
-  heading,
-  variant = "neutral",
-  live,
-  children,
-  ...props
-}: ComponentPropsWithoutRef<"div"> & {
-  heading?: ReactNode;
-  live?: "assertive" | "polite" | "off";
-  variant?: "error" | "warning" | "info" | "neutral";
-}) {
-  return (
-    <div
-      className={cn(
-        "rounded-md border p-3 text-sm",
-        variant === "error" &&
-          "border-red-200 bg-red-50 text-red-800 dark:border-red-800 dark:bg-red-900/20 dark:text-red-200",
-        variant === "warning" &&
-          "border-amber-200 bg-amber-50 text-amber-800 dark:border-amber-800 dark:bg-amber-900/20 dark:text-amber-200",
-        variant === "info" &&
-          "border-blue-200 bg-blue-50 text-blue-800 dark:border-blue-800 dark:bg-blue-900/20 dark:text-blue-200",
-        variant === "neutral" &&
-          "border-zinc-200 bg-zinc-50 text-zinc-700 dark:border-zinc-800 dark:bg-zinc-900/60 dark:text-zinc-300",
-        className
-      )}
-      role={live === "assertive" ? "alert" : "status"}
-      aria-live={live}
-      {...props}
-    >
-      {heading ? <p className="font-semibold">{heading}</p> : null}
-      {children}
-    </div>
-  );
-}
-
-export function LoginSpinner({
-  className,
-  ...props
-}: ComponentPropsWithoutRef<"div">) {
-  return (
-    <div
-      className={cn(
-        "h-5 w-5 animate-spin rounded-full border-2 border-zinc-300 border-t-zinc-950 dark:border-zinc-700 dark:border-t-zinc-50",
-        className
-      )}
-      role="status"
-      {...props}
-    />
-  );
-}
