+++ src/ui/index.ts
@@ -2,5 +2,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 export * from "./primitives";
+export * from "./appShell";
+export * from "./searchableControls";
 export * from "./styles";
 export { cn } from "@/lib/utils";
