+++ tests/admin-migration-boundary.test.ts
@@ -65,7 +65,6 @@ const forbiddenTextMarkers = [
 
 const oldComponentWrapperPaths = new Set([
   "src/components/alert",
-  "src/components/avatar",
   "src/components/badge",
   "src/components/button",
   "src/components/checkbox",
